% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auto_Complete.R
\name{auto_complete_var}
\alias{auto_complete_var}
\title{Auto-Complete}
\usage{
auto_complete_var()
}
\description{
An add-in function used to help auto-complete variables in the global environment.
}
\examples{
apple <- 10
# Type "ap" and run the add-in function

}
