% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auto_Surround.R
\name{auto_surround}
\alias{auto_surround}
\title{Auto-Surround}
\usage{
auto_surround(char)
}
\arguments{
\item{char}{The character to auto-surround with (character scalar).}
}
\value{
The output range, invisibly.
}
\description{
An add-in function used to help auto-surround text with give characters.
}
\examples{
# Type "apple, banana, cherry", put your cursor anywhere in a word (including on the edges),
#   & run the add-in function

}
