% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binned_Plots.R
\name{binned_one_way_data}
\alias{binned_one_way_data}
\title{Binned One-Way Data}
\usage{
binned_one_way_data(x, yData, weight = rep(1, length(x)),
  scaleWeight = TRUE, type = "quantile", bins = 10)
}
\arguments{
\item{x}{The vector to perform binning by.
If numeric, it will be binned according to \code{type}.
If non-numeric, it will be binned according to all the unique values.}

\item{yData}{A data.frame-like object (or numeric vector) whose numerical columns will be used
to calculate average values by bin.
Must have the same number of rows (or length) as the length of x.}

\item{weight}{An (optional) vector of weights for weighted binning,
and for taking weighted averages.
Must be the same length as x.}

\item{scaleWeight}{If \code{TRUE}, the weights will be scaled from 0 to 1 (i.e. as a percent).}

\item{type}{The type of binning desired.
Valid options are:
  \code{r == raw}, which will create a bin for each distinct value in \code{x}.
  \code{e == equal}, which will do the same as \code{p == pretty}
  \code{p == pretty}, which will cut the range of \code{x} into equally sized bins,
    with the endpoints of each bin being chosen to be nice round numbers.
  \code{q == quantile}, which will cut \code{x} into bins according to weighted quantiles.
  \code{m == min == minimum}, which will try to ensure there are a minimum of
    \code{bins} observations in each bin.}

\item{bins}{The number of bins to create (or, if \code{type} is in \code{m == min == minimum},
the desired minimum number of observations per bin).}
}
\description{
Calculate the average value of
}
