% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bounding_and_Rounding.R
\name{bounding}
\alias{bounding}
\alias{bound}
\alias{divide_round_bound}
\title{Bound or Divide/Round/Bound}
\usage{
bound(num, lower = -Inf, upper = Inf)

divide_round_bound(num, div = 1, lower = -Inf, upper = Inf, type = "round")
}
\arguments{
\item{num}{Numeric vector to bound.}

\item{lower}{Lower bound to floor the result at.}

\item{upper}{Upper bound to cap the result at.}

\item{div}{Numeric vector to divide \code{num} by.}

\item{type}{Type of rounding to be performed (one of "floor", "ceiling", or "round").}
}
\value{
\code{bound}: bounded numeric vector.

\code{divide_round_bound}: divided, rounded, and bounded integer vector.
}
\description{
Bound a number by two other numbers, or divide, round, and bound
}
\examples{
bound(1:10, 2, 8)
bound(runif(10), 0.25, 0.75)
bound(1:5, upper = 3)
bound(1:5, lower = 3)

divide_round_bound(1:10, 2, 2, 4)
divide_round_bound(runif(10), 0.01, 20, 80)
divide_round_bound(1:100, 10, type = "ceiling")
divide_round_bound(1:100, 10, type = "floor")
divide_round_bound(1:100, 10, type = "round")
divide_round_bound(1:100, 10)

}
