% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Dictionary.R
\name{$.columnTables}
\alias{$.columnTables}
\title{Extract Elements from a \code{columnTables} Object}
\usage{
\method{$}{columnTables}(colTables, col)
}
\arguments{
\item{colTables}{A \code{columnTables} object.}

\item{col}{The column to extract the table for (character scalar).}
}
\value{
A \code{table} of the desired column.
}
\description{
Extract Elements from a \code{columnTables} Object
}
\examples{
dd <- dataDict(mtcars)
dd$colTables$mpg

}
