% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Printing.R
\name{cat0}
\alias{cat0}
\title{\code{cat} Helper Function}
\usage{
cat0(...)
}
\arguments{
\item{...}{Arguments to be passed on to \code{sep}.}
}
\description{
\code{cat} Helper Function
}
\examples{
cat(1:10)
cat0(1:10)

}
