% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Add-Ins.R
\name{clean_code}
\alias{clean_code}
\title{Clean & Reformat Code}
\usage{
clean_code(code)
}
\arguments{
\item{code}{The code to be cleaned (character vector).}
}
\value{
The cleaned code (character vector).
}
\description{
Clean & Reformat Code
}
\examples{
KO:::clean_code("apple=5;banana<-1;cherry==100;1+2;10+20;1-2;1*2;1/2;1^2;1|2;1&2")

}
