% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cleaning.R
\name{clean_col_names}
\alias{clean_col_names}
\alias{clean_col_names.data.frame}
\alias{clean_col_names.data.table}
\title{Clean Column Names}
\usage{
clean_col_names(object, ...)

\method{clean_col_names}{data.frame}(object, short = FALSE, ...)

\method{clean_col_names}{data.table}(object, short = FALSE, ...)
}
\arguments{
\item{object}{The object to clean names for.}

\item{...}{further arguments passed to or from other methods.}

\item{short}{Whether to shorten certain cleaned name replacements (logical scalar).}
}
\value{
\code{clean_col_names.data.frame}:

\code{clean_col_names.data.table}:
}
\description{
Clean the column names of a \code{data.frame} or \code{data.table}.
}
\examples{
DF <- data.frame(
  "a[b.c]d" = 1:26, "e\\\\f;g1h" = letters,
  "i/j k_l+m-n.o(p)q_$_\%_kl" = sample(c(TRUE, FALSE), 26, replace = TRUE),
  check.names = FALSE
)
colnames(DF)
print(methods("clean_col_names"))
DF <- clean_col_names(DF, TRUE)
colnames(DF)

DT <- data.table::data.table(
  "a[b.c]d" = 1:26, "e\\\\f;g1h" = letters,
  "i/j k_l+m-n.o(p)q_$_\%_kl" = sample(c(TRUE, FALSE), 26, replace = TRUE)
)
colnames(DT)
clean_col_names(DT, TRUE)
colnames(DT)

}
