% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cleaning.R
\name{clean_names}
\alias{clean_names}
\title{Clean Names}
\usage{
clean_names(vec, short = FALSE)
}
\arguments{
\item{vec}{A character vector of names to clean.}

\item{short}{Whether to shorten certain cleaned name replacements (logical scalar).}
}
\value{
A character vector of cleaned names.
}
\description{
Clean names so that quotes (or backquotes) will not be necessary when subsetting the object.\cr
The names input will be cleaned by replacing all occurrences of:
\itemize{
  \item \code{$} with \code{"Dollar"} (or \code{"Dol"}, if \code{short == TRUE})
  \item \code{\%} with \code{"Percent"} (or \code{"Perc"}, if \code{short == TRUE})
  \item Any other non-word character (i.e. character not in \code{[a-z, A-Z, 0-9, _]}) with \code{"_"}
}
}
\examples{
clean_names(c("apple", "banana", "ab.cd", "ef_gh", "ij/kl", "a_\%_b", "a_$_b", "mn__op"))
clean_names(c("a_\%_b", "a_$_b"), short = FALSE)

}
