% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean_Code.R
\name{clean_selection}
\alias{clean_selection}
\title{Clean the Current Code Selection}
\usage{
clean_selection()
}
\value{
A list of the old/original code, and the new/cleaned code.
}
\description{
If no code is selected, the current line will be cleaned.
}
\examples{
# Highlight a line of code with spacing that would cause RStudio's style diagnostics to complain,
#   and run the add-in function

}
