% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Dictionary.R
\name{columnTables}
\alias{columnTables}
\title{Create a \code{columnTables} Object}
\usage{
columnTables(dict, df, table)
}
\arguments{
\item{dict}{A \code{dataDict} object.}

\item{df}{The data.frame-like object to calculate information for (should be a name, not an expression).}

\item{table}{What mode to use for tabulating each column (see \code{details} for more) (character/logical scalar).}
}
\value{

}
\description{
Create an object that is used to store the results of tabulating columns in a data.frame-like object.
This is currently only called from with a \code{dataDict} object.
}
\examples{

}
