% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compare_ggplots.R
\name{compare_ggplots}
\alias{compare_ggplots}
\title{Compare two ggplot objects}
\usage{
compare_ggplots(plot1, plot2, useTempFiles = TRUE)
}
\arguments{
\item{plot1}{First ggplot object to compare.}

\item{plot2}{Second ggplot object to compare.}

\item{useTempFiles}{Whether or not to use temporary files in the comparison.
Using temp files is faster, but some may wish to compare purely in memory.}
}
\value{
A boolean scalar showing whether the two plots are visually identical or not.
}
\description{
Compare two ggplot objects
}
\examples{
plot1 <- ggplot2::qplot(mpg, wt, data = mtcars, colour = cyl)
plot2 <- ggplot2::qplot(mpg, wt, data = mtcars, colour = cyl)
plot3 <- ggplot2::qplot(mpg, wt, data = mtcars, colour = carb)
compare_ggplots(plot1, plot2)
compare_ggplots(plot1, plot2, useTempFiles = FALSE)
compare_ggplots(plot1, plot3)
compare_ggplots(plot1, plot3, useTempFiles = FALSE)
}
