% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summarize.R
\name{count_distinct}
\alias{count_distinct}
\title{Count Distinct Values}
\usage{
count_distinct(data, use_base = TRUE)
}
\arguments{
\item{data}{A data frame/matrix/vector.}

\item{use_base}{Whether to use base \code{R} functions (i.e. \code{length/distinct})
or \code{dplyr::n_distinct}, which seems to have some performance bugs.}
}
\value{
Integer vector, named for data frames or matrixes with column names.
}
\description{
Count the number of distinct values in a data frame/matrix/vector.
If a data frame or matrix is passed, counts the number of distinct values by column.
If a vector is passed, just counts the number of distinct values.
}
\examples{
count_distinct(mtcars)
count_distinct(as.matrix(mtcars))
count_distinct(mtcars$mpg)

}
