% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Dictionary.R
\name{dim.dataDict}
\alias{dim.dataDict}
\title{Dimensions of a \code{dataDict}}
\usage{
\method{dim}{dataDict}(x)
}
\arguments{
\item{x}{A \code{dataDict} object.}
}
\value{
The dimensions of the original data.frame-like object.
}
\description{
Generic function to retrieve the dimensions of the data.frame-like object
  that the \code{dataDict} is based on.
`ncol` and `nrow` will also work, since they call `dim` in their implementation.
}
\examples{
dd <- dataDict(mtcars)
dim(dd)
ncol(dd)
nrow(dd)

}
