% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Dictionary.R
\name{dimension_names}
\alias{dimension_names}
\alias{dimnames.dataDict}
\alias{names.dataDict}
\title{Dimension Names of a \code{dataDict}}
\usage{
\method{dimnames}{dataDict}(x)

\method{names}{dataDict}(x)
}
\arguments{
\item{x}{A \code{dataDict} object.}
}
\value{
\code{dimnames}: The dimension names of the original data.frame-like object.

\code{names}: The column names of the original data.frame-like object.
}
\description{
Generic function to retrieve the dimension names of the data.frame-like object
  that the \code{dataDict} is based on.
`colnames` and `rownames` will also work, since they call `dimnames` in their implementation.
}
\examples{
dd <- dataDict(mtcars)
dimnames(dd)
colnames(dd) # Same as names(dd)
rownames(dd)
names(dd) # Same as colnames(dd)

}
