% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Dictionary.R
\name{dimensions}
\alias{dimensions}
\alias{dim.dataDict}
\title{Dimensions of a \code{dataDict}}
\usage{
\method{dim}{dataDict}(dict)
}
\arguments{
\item{dict}{A \code{dataDict} object.}
}
\value{
\code{dim}: The dimensions.
}
\description{
Generic functions to retrieve the dimensions of the data.frame-like object
  that the \code{dataDict} is based on.
}
\examples{
dd <- dataDict(mtcars)
dim(dd)
ncol(dd)
nrow(dd)

}
