% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Text_Removal.R
\name{find_pos}
\alias{find_pos}
\title{Find the Last Position}
\usage{
find_pos(txt)
}
\arguments{
\item{txt}{The text to search in (character scalar).}
}
\value{
The last position of the desired patterns.
}
\description{
Find the last position of the desired patterns.
}
\examples{
KO:::find_pos("mtcars \%>\% select(mpg, disp, cyl)")

}
