% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format_Functions.R
\name{format_percent}
\alias{format_percent}
\title{Format a numerical vector as percents}
\usage{
format_percent(x, accuracy = 0.01, vecNames = names(x))
}
\arguments{
\item{x}{Object to format as a percent}

\item{accuracy}{The number to round to the nearest multiple of
(e.g 3.14159 will round to 314.16% by default).
The absolutely value of the value passed will be used.
The rounding will be done after conversion to a percent.}

\item{vecNames}{Vector used to name the output}
}
\value{
Character vector of items formatted as percent.
}
\description{
Format a numerical vector as percents
}
\examples{
format_percent(0.12345)
format_percent(pi)
format_percent(0.1)

}
