% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Dictionary.R
\name{getTables}
\alias{getTables}
\title{Extract Tabuated Columns from a \code{dataDict}}
\usage{
getTables(dict, cols = NULL)
}
\arguments{
\item{cols}{If \code{elem == "colTables"}, the column(s) to extract.}

\item{elem}{The element to extract (character scalar).}
}
\value{
A named list of tables, corresponding to the requested columns.
}
\description{
Extract Tabuated Columns from a \code{dataDict}
}
\examples{
dd <- dataDict(mtcars)
getTables(dd)
getTables(dd, cols = NA)
getTables(dd, cols = NA)
getTables(dd, cols = "mpg")
getTables(dd, cols = c("mpg", "cyl"))

}
