% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation_Functions.R
\name{gini_weighted_normalized}
\alias{gini_weighted_normalized}
\title{Normalized Weighted Gini}
\usage{
gini_weighted_normalized(solution, weights = NULL, submission)
}
\arguments{
\item{solution}{Numerical vector of actual response.}

\item{weights}{Weights to assign to each prediction.}

\item{submission}{Predictions to score against the solution.}
}
\value{
Normalized weighted Gini index.
}
\description{
Calculate the normalized weighted Gini index of predictions against the solutions.
}
