% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inline_If.R
\name{iif}
\alias{iif}
\title{Inline If Statement}
\usage{
iif(condition, true, false)
}
\arguments{
\item{condition}{Condition to check (logical).}

\item{true}{Value to return when the condition is \code{TRUE}.}

\item{false}{Value to return when the condition is \code{FALSE}.}
}
\value{
Either the \code{true} or \code{false} parameter.
}
\description{
Inline if statement that checks the condition,
  and returns either the \code{true} or \code{false} paramater.
Similar to \code{ifelse}, but returns the entire value of the parameters.
}
\examples{
iif(TRUE, "a", "b")
iif(10 < 20, TRUE, FALSE)
print(iif(TRUE, "true", "false"))

# Compare iif to ifelse
iif(TRUE, 1:2, 3:4)
ifelse(TRUE, 1:2, 3:4)

}
