% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colored_Output.R
\name{info}
\alias{info}
\title{Print out Some Information}
\usage{
info(..., useMsg = TRUE, printOut = TRUE)
}
\arguments{
\item{...}{Text to print out (can also be named arguments to \code{message}/\code{cat},
e.g. \code{sep}, which defaults to " ").}

\item{useMsg}{Whether to use \code{message} to print the text (uses \code{cat} otherwise).}

\item{printOut}{Whether to print the value out (might just want the colored string raw).}
}
\value{
The colored string, invisibly.
}
\description{
Used to print out some information during the execution of a function.
This function prints out the text in green (#00FF00) using the \code{crayon} package.
}
\examples{
info("hello world")

}
