% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatting.R
\name{latex_table}
\alias{latex_table}
\title{Format a data frame in LaTeX code}
\usage{
latex_table(data)
}
\arguments{
\item{data}{Data frame to be formatted into a LaTeX table}
}
\value{
String to use used for printing a table in LaTeX.
}
\description{
Format a data frame to print out as a LaTeX table with side borders.
}
\examples{
latex_table(mtcars)

}
