% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summarize.R
\name{mode_stat}
\alias{mode_stat}
\title{Statistical Mode}
\usage{
mode_stat(vec, return_all_modes = FALSE, ignoreBlanks = FALSE)
}
\arguments{
\item{vec}{Atomic vector.}

\item{return_all_modes}{Whether or not to return a vector of all modes (for ties) (logical scalar).}

\item{ignoreBlanks}{Whether or not to ignore blanks (i.e. NA's and empty strings ("")) (logical scalar).}
}
\value{
Scalar of same type as \code{vec},
  or a vector of the same type if there are multiple modes,
  and \code{return_all_modes} is \code{TRUE}.
}
\description{
Ignoring NAs, find the most common element in a vector (statistical mode).
If multiple elements are tied for the mode,
  the last element in sorted order will be returned,
  unless all elements are specifically requested,
  in which case a vector of all modes will be returned.
}
\examples{
mode_stat(c(1, 1, 1, 2, 2))
mode_stat(c("a", "b", "b", "b", "c"))
mode_stat(c("a", "b"))
mode_stat(c("a", "b"), return_all_modes = TRUE)

}
