% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multi_Value_Columns.R
\name{multi_val_cols}
\alias{multi_val_cols}
\title{Extract Multi-Valued Columns}
\usage{
multi_val_cols(df)
}
\arguments{
\item{df}{An object inheriting from a \code{data.frame} (e.g. \code{data.table}/\code{tibble}).}
}
\value{
The columns of \code{df} that have more than one unique value.
}
\description{
Extract columns from a data frame that have more than one unique value.
}
\examples{
multi_val_cols(data.frame(a = c(1, 1, 1), b = c("a", "a", "a")))
multi_val_cols(data.frame(a = c(1, 1, 1), b = c(1, 2, 3),
                          c = c('a', 'a', 'a'), d = c('a', 'a', 'b')))
multi_val_cols(data.frame(a = c(1, 1, 3), b = c(1, 2, 3),
                          c = c('a', 'b', 'a'), d = c('a', 'a', 'b')))

}
