% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Dictionary.R
\name{names.columnTables}
\alias{names.columnTables}
\title{Get the Names of a \code{columnTables} Object}
\usage{
\method{names}{columnTables}(x)
}
\arguments{
\item{x}{A \code{columnTables} object.}
}
\value{
The names of the columns that have been tabulated already (character vector).
}
\description{
Get the names of the columns from a \code{columnTables} object that have
  been tabulated already (character vector).
}
\examples{
dd <- dataDict(mtcars)
names(dd$colTables)

}
