% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Notify.R
\name{notify}
\alias{notify}
\title{Send a Toast Notification}
\usage{
notify(
  msg,
  title = "",
  img = "C:/Program Files/RStudio/resources/tutorial_resources/rstudio-logo.png"
)
}
\arguments{
\item{msg, title, img}{The notification message/title/image path (character scalars).}
}
\description{
Send a toast notification on Windows.
}
\examples{
notify("Hello", "World")

}
