% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Missing_Values.R
\name{num_missing}
\alias{num_missing}
\title{Find the number of missing values}
\usage{
num_missing(vec)
}
\arguments{
\item{vec}{Vector to count missing values in.}
}
\value{
Numeric scalar containing the number of missing values.
}
\description{
Count the number of missing values in a vector.
Counts the number of NA values for non-character vectors,
  and the number of NA or "" values for characters.
}
\examples{
num_missing(c(NA))
num_missing(c(1, 2, 3, 4, 5))
num_missing(c(1, 2, NA, 4, 5))

}
