% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatting.R
\name{pad_vector}
\alias{pad_vector}
\title{Convert a vector to a padded character column}
\usage{
pad_vector(vec, header = NA, padding = " ", alignment = "R")
}
\arguments{
\item{vec}{Vector to pad values of}

\item{header}{Optional name of vector to add to top of return value}

\item{padding}{Character to pad output vector with}

\item{alignment}{Alignment of text in output (one of 'R', 'L', 'C')}
}
\value{
A character vector padded with spaces to give uniform width.
}
\description{
Convert a vector to a padded character column
}
\examples{
pad_vector(c(1, 2, 300))
pad_vector(c(1, 2), header = "Header")
pad_vector(c(1, 2, 300), padding = "_")
pad_vector(c(1, 200, 3), alignment = "L")
pad_vector(c(100, 2, 30), alignment = "C")

}
