% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Missing_Values.R
\name{percent_missing_vals}
\alias{percent_missing_vals}
\title{Find percent of missing values}
\usage{
percent_missing_vals(vec)
}
\arguments{
\item{vec}{Vector to count missing values in}
}
\value{
A number containing the percent of missing values.
}
\description{
Count the percent of missing values in a vector.
Counts the percent of NA values for non-character vectors,
  and the number of NA or "" values for characters.
}
\examples{
percent_missing_vals(c(NA))
percent_missing_vals(c(1, 2, 3, 4, 5))
percent_missing_vals(c(1, 2, NA, 4, 5))

}
