% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{Pipes}
\alias{Pipes}
\title{Pipe Operators}
\description{
See \code{magrittr::\link[magrittr]{\%$\%}}
  or \code{magrittr::\link[magrittr]{\%<>\%}}
  or \code{magrittr::\link[magrittr]{\%>\%}}
  or \code{magrittr::\link[magrittr]{\%T>\%}} for details.
}
\keyword{internal}
