% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipe_Backspace.R
\name{pipe_backspace}
\alias{pipe_backspace}
\title{Pipe Backspace}
\usage{
pipe_backspace()
}
\value{
The deleted code (character scalar).
}
\description{
An add-in function that deletes the code between the cursor and the last pipe.
}
\examples{
# Type "1:10 \%>\% mean()", put your cursor anywhere after the pipe, & run the add-in function

}
