% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Dictionary.R
\name{print.colTables}
\alias{print.colTables}
\title{Print a \code{dataDict} Object}
\usage{
\method{print}{colTables}(colTables)
}
\arguments{
\item{colTables}{A \code{columnTables} object.}
}
\value{
The summary of \code{colTables}, invisibly (i.e. \code{summary(colTables)}).
}
\description{
Print a \code{dataDict} Object
}
\examples{
dd <- dataDict(mtcars)
dd$colTables
print(dd$colTables)

}
