% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatting.R
\name{pvalue_stars}
\alias{pvalue_stars}
\title{P-Value Stars}
\usage{
pvalue_stars(p_values, alignment = "R")
}
\arguments{
\item{p_values}{A numeric vector of p-values.}

\item{alignment}{Which side to align the significance stars to
(padding spaces will be added to the other side).}
}
\value{
A character vector of the p-value significance stars.
}
\description{
Calculate the significance stars for p-values.
}
\examples{
pvalue_stars(c(-Inf, 0, 0.0005, 0.001, 0.005, 0.01, 0.025, 0.05, 0.1, 0.15, 1, Inf, NA))
pvalue_stars(c(-Inf, 0, 0.001, 0.01, 0.025, 0.05, 0.1, 0.15, 1, Inf, NA), alignment = "L")

}
