% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation_Functions.R
\name{quasi_poisson_aic_bic}
\alias{quasi_poisson_aic_bic}
\title{Calculate AIC & BIC}
\usage{
quasi_poisson_aic_bic(model)
}
\arguments{
\item{model}{Quasi-Poisson GLM to calculate the AIC/BIC for.}
}
\value{
A list of two elements, AIC and BIC.
}
\description{
Calculate the AIC and BIC for a quasi-Poisson GLM by calculating training a normal Poisson GLM.
}
\examples{
# quasi_poisson_aic_bic(quasiPoissonGLM)

}
