% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Missing_Values.R
\name{replace_missing}
\alias{replace_missing}
\title{Replace missing values in a vector}
\usage{
replace_missing(vec, method = "mode", with, return_imputed = FALSE)
}
\arguments{
\item{vec}{Vector in which to replace missing values.}

\item{method}{The method to use to determine the imputed value.}

\item{with}{The value to impute missing values with (if specified, method will be ignored).}

\item{return_imputed}{Whether or not to return the value that missing elements were imputed with.}
}
\value{
Vector with missing values replaced as desired,
  or a list of that and the replacement (imputed) value.
}
\description{
Replace missing values in a vector with either the mode, median,
  or mean of the vector, removing NA values in the calculation.
}
\examples{
replace_missing(c(1, 1, 1, 2, NA))
replace_missing(c(1, 1, 1, 2, NA), method = "mode") # Same as above
replace_missing(c(1, NA, 1, 2), method = "median")
replace_missing(c(1, NA, 1, 2), method = "mean")
replace_missing(c(1, NA, 1, 2), method = "mean", return_imputed = TRUE)
replace_missing(c(1, NA, 1, 2), with = 5)
replace_missing(c(1, NA, 1, 2), method = "mean", with = 5)

}
