% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Console_Add-Ins.R
\name{run_chunk}
\alias{run_chunk}
\title{Run Code in the Console}
\usage{
run_chunk()
}
\value{
The code in the console, invisibly (character scalar).
}
\description{
Run the code currently in the console, without clearing it or moving the selection.
}
\examples{
# Enter any code in the console, & run the add-in function

}
