% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_Console.R
\name{run_console}
\alias{run_console}
\title{Run Code in the Console}
\usage{
run_console()
}
\value{
The code in the console, invisibly (character scalar).
}
\description{
Run the code currently in the console, without clearing it or moving the selection.
}
\examples{
# Enter any code in the console, & run the add-in function

}
