% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Console_Add-Ins.R
\name{run_fn}
\alias{run_fn}
\title{Run a Function}
\usage{
run_fn()
}
\description{
Run a user-set function in the background at the press of a button.
If the function has not been set, this function will print out some code to help the user get started.
Once the function has been set, running this add-in will execute the set function.
}
\examples{
# To make use of this add-in, run it (preferably by assigning a shortcut),
#   and it will print run a commented out line which will allow you to set a function.
# Once the `.KO_fn` object exists, running this add-in will execute the set function.

.KO_fn <- function() mean(1:10)
.KO_fn
.KO_fn()

}
