% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Console_Add-Ins.R
\name{run_selection}
\alias{run_selection}
\title{Run Selected Code}
\usage{
run_selection()
}
\value{
The selected code, invisibly (character scalar).
}
\description{
Run the code currently selected, without clearing the console or moving the selection.
}
\examples{
# Select any code, & run the add-in function

}
