% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Add-Ins.R
\name{save_value_to_clipboard_variable}
\alias{save_value_to_clipboard_variable}
\title{Save Value to Clipboard Variable}
\usage{
save_value_to_clipboard_variable()
}
\value{
The value saved to the clipboard variable, invisibly.
}
\description{
Save the value of the currently selected expression to a variable.
This is particularly valuable in RStudio Server,
  where the clipboard is not available programatically.
}
\examples{
# Select an expression that can be evaluated to a value, & run this add-in

}
