% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Console_Add-Ins.R
\name{send_selection_to_console}
\alias{send_selection_to_console}
\title{Send Code to the Console}
\usage{
send_selection_to_console()
}
\value{
The inserted text, invisibly (character scalar).
}
\description{
Insert the currently selected code into the console.
}
\examples{
# Highlight some code in the source pane, & run the add-in function

}
