% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Add-Ins.R
\name{server-clipboard}
\alias{server-clipboard}
\alias{send_value_to_source_editor}
\alias{paste_value_from_clipboard_variable}
\title{Send Value to Source Editor}
\usage{
send_value_to_source_editor()

paste_value_from_clipboard_variable()
}
\value{
The value of the currently selected expression, invisibly.

The value read from the clipboard variable, invisibly.
}
\description{
Send the value of the currently selected expression to the source editor.
This is particularly valuable in RStudio Server,
  where the clipboard is not available programatically.

Paste the value of the clipboard variable to the active context.
This is particularly valuable in RStudio Server,
  where the clipboard is not available programatically.
}
\examples{
# Select an expression that can be evaluated to a value, & run this add-in

# After running `save_value_to_clipboard_variable` on a valid expression, run this add-in

}
