% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snippet_Pipe_Check.R
\name{snippet_pipe_check}
\alias{snippet_pipe_check}
\title{Snippet Pipe Check}
\usage{
snippet_pipe_check(pipeStr = "\%>\% ", afterStr = "", elseStr = "")
}
\arguments{
\item{pipeStr}{The pipe to insert if needed (character scalar).}

\item{afterStr}{Text to insert after the pipe (character scalar).}

\item{elseStr}{Text to insert if no pipe is needed (character scalar).}
}
\value{
The deleted code (character scalar).
}
\description{
An add-in function that decides whether or not to insert a pipe as well during snippet insertion.
}
\examples{


}
