% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Dictionary.R
\name{[.columnTables}
\alias{[.columnTables}
\title{Extract Elements from a \code{columnTables} Object}
\usage{
\method{[}{columnTables}(colTables, cols = NULL)
}
\arguments{
\item{colTables}{A \code{columnTables} object.}

\item{cols}{The columns to extract the tables for (character vector).}
}
\value{
A list of \code{table}s of the desired columns.
}
\description{
Extract Elements from a \code{columnTables} Object
}
\examples{
dd <- dataDict(mtcars)
dd$colTables["mpg"]
dd$colTables[c("mpg", "cyl", "am")]

}
