% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Dictionary.R
\name{[.dataDict}
\alias{[.dataDict}
\title{Extract Elements from a \code{dataDict} Object}
\usage{
\method{[}{dataDict}(dict, elem, col = NULL)
}
\arguments{
\item{dict}{A \code{dataDict} object.}

\item{elem}{The element to extract (character scalar).}

\item{col}{If \code{elem == "colTables"}, the column(s) to extract.}
}
\value{
The desired element extracted from the \code{dataDict} object.
}
\description{
Extract Elements from a \code{dataDict} Object
}
\examples{
dd <- dataDict(mtcars)
dd["ncol"]
dd["numUnique"]
dd["colTables"]
dd["colTables", NA]
dd["colTables", "mpg"]
dd["colTables", c("mpg", "cyl")]

dd2 <- dataDict(mtcars, table = TRUE, lazyTable = FALSE)
dd2["colTables", NA]
dd2["colTables", "mpg"]
dd2["colTables", c("mpg", "cyl")]

}
