% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summarize.R
\name{summarize}
\alias{summarize}
\alias{summarize_vec}
\alias{summarize_df}
\title{Summarize a Data Frame/Matrix/Vector}
\usage{
summarize_vec(vec, as_vec = FALSE, get_table = FALSE, useNA = "ifany")

summarize_df(data, as_vec = FALSE, get_table = FALSE, useNA = "ifany")
}
\arguments{
\item{vec}{Atomic vector.}

\item{as_vec}{Whether to return the values as a named vector instead of a list (logical).}

\item{get_table}{Whether to tabulate the vector as well.}

\item{useNA}{Value to pass to table to determine whether or not to use NA values
(one of "no", "ifany", and "always").}

\item{data}{A data frame/matrix/vector.}
}
\value{
List of summary statistics.

List of lists of summary statistics.
}
\description{
Takes in a data frame/matrix/vector and calculates summary statistics based on the type.
If \code{get_table} is \code{TRUE}, returns a table of counts as well.
}
\examples{
summarize_vec(mtcars$mpg)
summarize_vec(mtcars$mpg, as_vec = TRUE)
summarize_vec(mtcars$mpg, get_table = TRUE)
summarize_vec(rownames(mtcars))
summarize_vec(c(1, 1, 1, 2, 2, NA), get_table = TRUE)
summarize_vec(c(1, 1, 1, 2, 2, NA), get_table = TRUE, useNA = "no")
summarize_vec(sample(c(TRUE, FALSE), 10, replace = TRUE))
summarize_vec(sample(c(TRUE, FALSE), 10, replace = TRUE), as_vec = TRUE)

summarize_df(mtcars)
summarize_df(mtcars, as_vec = TRUE)
summarize_df(mtcars, get_table = TRUE)

}
