% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Dictionary.R
\name{summary.colTables}
\alias{summary.colTables}
\title{Summarize a \code{dataDict} Object}
\usage{
\method{summary}{colTables}(colTables)
}
\arguments{
\item{colTables}{A \code{columnTables} object.}
}
\value{
The summary of \code{colTables}, invisibly (i.e. \code{summary(colTables)}).
}
\description{
Summarize a \code{dataDict} Object
}
\examples{
dd <- dataDict(mtcars)
summary(dd$colTables)

}
