% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Dictionary.R
\name{summary.columnTables}
\alias{summary.columnTables}
\title{Summarize a \code{dataDict} Object}
\usage{
\method{summary}{columnTables}(object, ...)
}
\arguments{
\item{object}{A \code{columnTables} object.}

\item{...}{Currently not used.}
}
\value{
A summary of \code{object}.
}
\description{
Summarize a \code{dataDict} Object
}
\examples{
dd <- dataDict(mtcars)
summary(dd$colTables)

}
