% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Dictionary.R
\name{summary.dataDict}
\alias{summary.dataDict}
\title{Summarize a \code{dataDict} Object}
\usage{
\method{summary}{dataDict}(object, ...)
}
\arguments{
\item{object}{A \code{dataDict} object.}

\item{...}{Currently not used.}
}
\value{
A summary of \code{object}.
}
\description{
Summarize a \code{dataDict} Object
}
\examples{
\dontrun{
  summary(dataDict(mtcars))
}

}
