% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summarize.R
\name{table_df}
\alias{table_df}
\title{Tabulate a Data Frame/Matrix}
\usage{
table_df(data, max_distinct = 250, group = TRUE, useNA = "ifany")
}
\arguments{
\item{data}{A data frame/matrix}

\item{max_distinct}{Maximum number of unique elements to return ungrouped (integer)}

\item{group}{Whether or not to group columns with more than \code{max_distinct} (logical)}

\item{useNA}{Value to pass to table to determine whether or not to use NA values
(one of "no", "ifany", and "always")}
}
\value{
List of tables.
}
\description{
Calls the table function on each column in a data frame/matrix,
  but only for columns with up to a certain number of unique values.
}
\examples{
table_df(mtcars)
table_df(as.matrix(mtcars))
table_df(mtcars$mpg)
table_df(c(1, 2, 3, NA, NA))
table_df(c(1, 2, 3, NA, NA), useNA = "no")
table_df(c(1, 2, 3, NA, NA), useNA = "always")

}
