% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auto_Complete.R
\name{tkDropDown}
\alias{tkDropDown}
\title{Create a TCL/TK Drop-Down List}
\usage{
tkDropDown(varList, title = "Select a Word")
}
\arguments{
\item{varList}{Variable names to choose from (character vector)/}

\item{title}{Title of the pop-up (character scalar).}
}
\value{
The selected variable name.
}
\description{
Create a TCL/TK Drop-Down List
}
\examples{
\dontrun{
  tkDropDown(letters)
  tkDropDown(letters, "Select a Variable")
}

}
