% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trimming.R
\name{trim_glm}
\alias{trim_glm}
\title{Remove excess parts of a GLM}
\usage{
trim_glm(model)
}
\arguments{
\item{model}{A GLM object}
}
\value{
A trimmed-down GLM, with only the parts needed for predicton.
}
\description{
Remove parts of a GLM not needed for prediction to trim down size.
}
\examples{
mtcarsGLM <- glm(mpg ~ ., data = mtcars)
object.size(mtcarsGLM)
object.size(trim_glm(mtcarsGLM))

}
