% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trimming.R
\name{trim_glm_summary}
\alias{trim_glm_summary}
\title{Remove excess parts of a GLM summary}
\usage{
trim_glm_summary(model_smy)
}
\arguments{
\item{model_smy}{Summary of a}
}
\value{
A trimmed-down GLM summary.
}
\description{
Remove parts of a GLM summary to shrink the size.
}
\examples{
mtcarsGLMSummary <- summary(glm(mpg ~ ., data = mtcars))
object.size(mtcarsGLMSummary)
object.size(trim_glm_summary(mtcarsGLMSummary))

}
