% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Dictionary.R
\name{updateDD}
\alias{updateDD}
\title{Update a \code{dataDict} Object}
\usage{
updateDD(dict, df)
}
\arguments{
\item{dict}{A \code{dataDict} object.}

\item{df}{The data.frame-like object to calculate information for.}
}
\description{
Update a \code{dataDict} Object
}
\examples{
df <- mtcars
dd <- dataDict(df)
dd$colTables$mpg
df$mpg <- df$mpg * 2
update(dd, df)
dd$colTables$mpg

}
