% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clipping.R
\name{vec_to_code}
\alias{vec_to_code}
\title{Transform a Vector to Code}
\usage{
vec_to_code(vec, width = 100, indent = 2, tol = 5)
}
\arguments{
\item{vec}{The vector to transform (atomic vector).}

\item{width}{The approximate desired line width (integer scalar).}

\item{indent}{The number of spaces to use as an indent (integer scalar).}

\item{tol}{The tolerance for going above thw width (integer scalar).}
}
\value{
A character scalar of the vector transformed to code.
}
\description{
Transform a Vector to Code
}
\examples{

set.seed(112)
s <- replicate(100, sample(letters, floor(runif(1, 8, 13)), replace = TRUE)
v <- vec_to_code(s)
cat(v)
purrr::map_int(strsplit(v, "\\n")[[1]], nchar)

}
