% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Printing.R
\name{makeOutputRows}
\alias{makeOutputRows}
\alias{vecPrint}
\title{Prepare the Rows of the Output}
\usage{
makeOutputRows(vec, numCols, maxChars)

vecPrint(vec, maxLen, printOut = TRUE)
}
\arguments{
\item{vec}{The (atomic) vector to print out.}

\item{maxLen}{The maximum length of a string to print out (integer scalar).}

\item{printOut}{Whether to print out the formatted output (logical scalar).}
}
\value{
The output rows vector.

The final output row vector.
}
\description{
Prints out a vector columnwise, using the full width available.
This is usually called for its side effects, but will also return the (unformatted) output rows vector.
}
\examples{

}
