% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auto_Complete.R
\name{auto_complete_var}
\alias{auto_complete_var}
\title{Auto-Complete}
\usage{
auto_complete_var()
}
\value{
The variable name that is inserted, invisibly (character scalar).
}
\description{
An add-in function used to help auto-complete variables in the global environment.
}
\examples{
apple <- 10
# Type "ap" & run the add-in function

var1 <- 1; var2 <- 2; var3 <- 3
# Type "var" & run the add-in function

# source('~/GitHub/KO/R/Auto_Complete.R'); auto_complete_var()

}
