% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colored_Output.R
\name{colored_output}
\alias{colored_output}
\alias{info}
\alias{color_list}
\title{Print out some Information}
\usage{
info(..., useMsg = TRUE, printOut = TRUE)

color_list(vec, cols = c("#FF00FF", "#FF8800", "#00FF00"),
  printOut = TRUE)
}
\arguments{
\item{...}{Text to print out (can also be named arguments to \code{message}/\code{cat},
e.g. \code{sep}, which defaults to " ").}

\item{useMsg}{Whether to use \code{message} to print the text (uses \code{cat} otherwise).}

\item{printOut}{Whether to print the value out (might just want the colored string raw).}

\item{vec}{Vector to print (preferably a list).}

\item{cols}{Colors to use (vector of three strings that \code{crayon::make_style} accepts).}
}
\value{
The colored string, invisibly.

The colored output lines, invisibly.
}
\description{
Used to print out some information during the execution of a function.
This function prints out the text in green (#00FF00) using the \code{crayon} package.

Used to print out a simple list, with color, and aligned values.
}
\examples{
info("hello world")

color_list(list(apple = "banana", cherry = "orange", peach = "pineapple"))

}
