% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compare_DF.R
\name{compare_dfs}
\alias{compare_dfs}
\title{Compare Data-Frames}
\usage{
compare_dfs(df1, df2, printColDiffs = TRUE, tol = 1e-10)
}
\arguments{
\item{df1}{The first object inheriting from a data.frame.}

\item{df2}{The second object inheriting from a data.frame.}

\item{printColDiffs}{Whether or not to print messages showing
the differences between columns (logical scalar).}

\item{tol}{The acceptable tolerance for equality between numeric columns (numeric scalar).}
}
\value{
A list of column comparison data for each non-identical column.
}
\description{
Compare two data-frame (like) objects.
The function will compare the objects by column names, and print information about differences.
}
\examples{

# Returns an empty list, since the objects are identical
compare_dfs(mtcars, mtcars)

# Prints a message about the difference in column names
compare_dfs(mtcars, mtcars[, -1])

# Prints a message about the difference in values, & returns a tibble showing the differences
compare_dfs(mtcars, dplyr::mutate(mtcars, mpg = mpg * 2))

# Same as the above, but doesn't print the message about the differences
compare_dfs(mtcars, dplyr::mutate(mtcars, mpg = mpg * 2), printColDiffs = FALSE)

}
