% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatting.R
\name{format_percent}
\alias{format_percent}
\title{Format a numerical vector as percents}
\usage{
format_percent(vec, accuracy = 0.01, vec_names = names(vec),
  silent = FALSE)
}
\arguments{
\item{vec}{Vector to format as a percent}

\item{accuracy}{The number to round to the nearest multiple of
(e.g 3.14159 will round to 314.16\% by default).
The absolute value of this parameter will be used.
The rounding will be done after conversion to a percent.}

\item{vec_names}{Vector used to name the output}

\item{silent}{If TRUE, will not print out a message if NA values are found}
}
\value{
Character vector of items formatted as percent.
}
\description{
Format a numerical vector as a character vector holding the values formatted as percents.
Will throw an error for non-numeric vetcors, and  print a message if NA values are found.
}
\examples{
format_percent(0.12345)
format_percent(pi)
format_percent(0.1)
format_percent(0.12345, accuracy = 0.1)
format_percent(c(0.12345, 0.54321), vec_names = c("a", "b"))

}
