% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation_Functions.R
\name{gini_weighted}
\alias{gini_weighted}
\title{Weighted Gini}
\usage{
gini_weighted(solution, predictions, weights = 1)
}
\arguments{
\item{solution}{Numerical vector of actual response.}

\item{predictions}{Predictions to score against the solution.}

\item{weights}{Weights to assign to each prediction.}
}
\value{
Weighted Gini index.
}
\description{
Calculate the weighted Gini index of predictions against the solutions.
}
