% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatting.R
\name{plain_text_table}
\alias{plain_text_table}
\title{Returns a data frame in plain text with padded (equal width) columns}
\usage{
plain_text_table(data, sep = " | ", cat_out = TRUE)
}
\arguments{
\item{data}{A data frame to convert to plain text}

\item{sep}{Characters to bs used as column seperators}

\item{cat_out}{Whether or not to print the value to standard output}
}
\value{
If cat_out, nothing, else a string containing the table in plain text form.
}
\description{
Returns a data frame in plain text with padded (equal width) columns
}
\examples{
plain_text_table(head(mtcars))
plain_text_table(head(cars))
plain_text_table(data.frame(a = 1:2, b = 3:4))
plain_text_table(head(mtcars), sep = "+")
plain_text_table(head(mtcars), cat_out = FALSE)

}
