% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Dictionary.R
\name{print.dataDict}
\alias{print.dataDict}
\title{Print a \code{dataDict} Object}
\usage{
\method{print}{dataDict}(dict)
}
\arguments{
\item{dict}{A \code{dataDict} object.}
}
\description{
Print a \code{dataDict} Object
}
\examples{
dataDict(mtcars)
print(dataDict(mtcars))
dd <- dataDict(mtcars)
print(dd)

}
