% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quasi_Poisson.R
\name{quasi_poisson}
\alias{quasi_poisson}
\title{Create a function to define AIC within a quasi Poisson family}
\usage{
quasi_poisson(...)
}
\arguments{
\item{...}{Input values to families}
}
\value{
Quasi Poisson family with AIC/BIC defined.
}
\description{
Create a function to define AIC within a quasi Poisson family
}
\examples{
\donttest{
  data(dataCar, package = "insuranceData")
  dataCarGLM <- glm(numclaims ~ veh_value + veh_age + gender + agecat,
                    data = dataCar, family = quasi_poisson, offset = log(exposure))
  dataCarGLMSummary <- summary(dataCarGLM)
}

}
