% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quasi_Poisson.R
\name{quasi_poisson_aic_bic}
\alias{quasi_poisson_aic_bic}
\title{Calculate AIC & BIC}
\usage{
quasi_poisson_aic_bic(model)
}
\arguments{
\item{model}{Quasi-Poisson GLM to calculate the AIC/BIC for}
}
\value{
A list of two elements, AIC and BIC.
}
\description{
Calculate the AIC and BIC for a quasi-Poisson GLM by training a normal Poisson GLM.
}
\examples{
\dontrun{
  data(dataCar, package = "insuranceData")
  dataCarGLM <- glm(numclaims ~ veh_value + veh_age + gender + agecat,
                    data = dataCar, family = quasi_poisson, offset = log(exposure))
  quasi_poisson_aic_bic(dataCarGLM)
}

}
