% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Text_Removal.R
\name{remove_to_text}
\alias{remove_to_text}
\alias{pipe_backspace}
\alias{comma_backspace}
\title{Remove to Text}
\usage{
remove_to_text(backTo, elseBackTo = "\\\\(|\\\\)|\\\\[|\\\\]",
  backward = TRUE)

pipe_backspace()

comma_backspace()
}
\arguments{
\item{backTo}{The regex pattern to remove text to (character scalar).}

\item{elseBackTo}{If \code{backTo} is not found before this regex pattern, only remove up to these (character scalar).}

\item{backward}{Whether to remove text before the cursor (default) or text after it (boolean scalar).}
}
\value{
The deleted code, invisibly (character scalar).
}
\description{
An add-in helper function that deletes the code between the cursor and the desired string.

An add-in function that deletes the code between the cursor and the last pipe.

An add-in function that deletes the code between the cursor and the last comma.
}
\examples{
# Type "mtcars \%>\% mutate(a = 1) \%>\% filter(cyl = 2)", & run the add-in function

# Type "c(apple, banana, cherry)", & run the add-in function

}
