% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compare_Sets.R
\name{set_comp}
\alias{set_comp}
\title{Set Comparison}
\usage{
set_comp(set1, set2, compNames = FALSE, printOut = TRUE)
}
\arguments{
\item{set1}{The first set to compare.}

\item{set2}{The second set to compare.}

\item{compNames}{Whether to compare the names of the sets, or the sets themselves (logical scalar).}

\item{printOut}{Whether to print out the differences (logical scalar).}
}
\value{
A two-element list of the set differences.
}
\description{
Compare two sets, printing out the differences if desired.
}
\examples{
set_comp(1:10, 6:15)
set_comp(mtcars[, -1], mtcars[, -2])
set_comp(c(a = 1, b = 2, c = 3), c(a = 1, b = 2, d = 4))
set_comp(c(a = 1, b = 2, c = 3), c(a = 1, b = 2, d = 4), compNames = FALSE)
set_comp(c(a = 1, b = 2, c = 3), c(a = 1, b = 2, d = 4), compNames = TRUE)
comp <- set_comp(c(a = 1, b = 2, c = 3), c(a = 1, b = 2, d = 4), printOut = FALSE)

}
