% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snippet_Pipe_Check.R
\name{snippet_pipe_check}
\alias{snippet_pipe_check}
\title{Snippet Pipe Check}
\usage{
snippet_pipe_check(pipeStr = "\%>\% ", afterStr = "", elseStr = "")
}
\arguments{
\item{pipeStr}{The pipe to insert if needed (character scalar).}

\item{afterStr}{Text to insert after the pipe (character scalar).}

\item{elseStr}{Text to insert if no pipe is needed (character scalar).}
}
\value{
The deleted code (character scalar).
}
\description{
Decides whether or not to insert a pipe as well during snippet insertion.
}
\examples{
# Set up a snippet as:
# snippet mu
# `r KO:::snippet_pipe_check();`mutate(${0})

# Then type "mtcars mu", & run the add-in function
# Or type "mtcars \%>\% mu", & run the add-in function
# The output for both will be "mtcars \%>\% mutate()"

}
