% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snippet_Pipe_Check.R
\name{snippet_pipe_check}
\alias{snippet_pipe_check}
\title{Snippet Pipe Check}
\usage{
snippet_pipe_check()
}
\value{
The deleted code (character scalar).
}
\description{
An add-in function that decides whether or not to insert a pipe as well during snippet insertion.
}
\examples{
# Type "1:10 \%>\% mean()", put your cursor anywhere after the pipe, & run the add-in function

}
