% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Dictionary.R
\name{[.dataDict}
\alias{[.dataDict}
\title{Extract Elements from a \code{dataDict} Object}
\usage{
\method{[}{dataDict}(dict, elem, cols = NULL)
}
\arguments{
\item{dict}{A \code{dataDict} object.}

\item{elem}{The element to extract (character scalar).}

\item{cols}{If \code{elem == "colTables"}, the column(s) to extract (character vector).}
}
\value{
The desired element extracted from the \code{dataDict} object.
}
\description{
Extract Elements from a \code{dataDict} Object
}
\examples{
dd <- dataDict(mtcars)
dd["numUnique"]
dd["colTables", "mpg"]
dd["colTables", NA]
dd["colTables"]

dd2 <- dataDict(mtcars, tableMode = TRUE)
dd2["colTables", "mpg"]
dd2["colTables", NA]

# The following lines work, but cause automated testing problems
# dd["colTables", c("mpg", "cyl")]
# dd2["colTables", c("mpg", "cyl")]

}
