% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Dictionary.R
\name{summary.colTables}
\alias{summary.colTables}
\title{Summarize a \code{dataDict} Object}
\usage{
\method{summary}{colTables}(colTables)
}
\arguments{
\item{colTables}{A \code{columnTables} object.}
}
\value{
A summary of \code{colTables}.
}
\description{
Summarize a \code{dataDict} Object
}
\examples{
\dontrun{
  dd <- dataDict(mtcars)
  summary(dd$colTables)
}

}
