% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Dictionary.R
\name{summary.dataDict}
\alias{summary.dataDict}
\title{Summarize a \code{dayaDict} Object}
\usage{
\method{summary}{dataDict}(dict)
}
\arguments{
\item{dict}{A \code{dataDict} object.}
}
\value{

}
\description{
Summarize a \code{dayaDict} Object
}
\examples{
summary(dataDict(mtcars))

}
