% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Dictionary.R
\name{summary.dataDict}
\alias{summary.dataDict}
\title{Summarize a \code{dataDict} Object}
\usage{
\method{summary}{dataDict}(dict)
}
\arguments{
\item{dict}{A \code{dataDict} object.}
}
\value{
A summary of \code{dict}.
}
\description{
Summarize a \code{dataDict} Object
}
\examples{
\dontrun{
  summary(dataDict(mtcars))
}

}
